pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "ergon-studio"
version = "0.1.0"
description = "OpenAI-compatible orchestration proxy for local coding models."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "agent-framework-core",
  "PyYAML",
]

[project.optional-dependencies]
dev = [
  "mypy",
  "pytest",
  "pytest-cov",
  "ruff",
  "types-PyYAML",
]

[project.scripts]
ergon-studio = "ergon_studio.proxy_cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
  "B",
  "E",
  "F",
  "I",
  "UP",
]

[tool.mypy]
python_version = "3.12"
files = ["ergon_studio"]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true
